Added addItem & clearItems, as well as basic Item schema
