+++ server/models/room.model.js
@@ -6,40 +6,6 @@ const Template = require("./template.model");
 let Room = {};
 
 /**
- * type RoomItem struct {
-	ID string `json:"id" rethinkdb:"id"`
-	Name string `json:"name" rethinkdb:"name"`
-	Quantity int `json:"quantity" rethinkdb:"quantity"`
-	ClaimedBy string `json:"claimedBy" rethinkdb:"claimedBy"`
-	VisibleInEditor bool `json:"visibleInEditor" rethinkdb:"visibleInEditor"`
-	Dimensions ItemDimensions `json:"dimensions" rethinkdb:"dimensions"`
-	EditorPosition EditorPoint `json:"editorPosition" rethinkdb:"editorPosition"`
-	EditorRotation float64 `json:"editorRotation" rethinkdb:"editorRotation"`
-	EditorLocked bool `json:"editorLocked" rethinkdb:"editorLocked"`
-	EditorZIndex float64 `json:"editorZIndex" rethinkdb:"editorZIndex"`
-}
-
-type ItemDimensions struct {
-	Width float64 `json:"width" rethinkdb:"width"`
-	Length float64 `json:"length" rethinkdb:"length"`
-	Height float64 `json:"height" rethinkdb:"height"`
-}
-
-type EditorPoint struct {
-	X float64 `json:"x" rethinkdb:"x"`
-	Y float64 `json:"y" rethinkdb:"y"`
-}
-
-type Room struct {
-	ID string `json:"id" rethinkdb:"id"`
-	Name string `json:"name" rethinkdb:"name"`
-	TemplateID string `json:"templateId" rethinkdb:"templateId"`
-	Items []RoomItem `json:"items" rethinkdb:"items"`
-	Vertices []EditorPoint `json:"vertices" rethinkdb:"vertices"`
-}
- */
-
-/**
  * Create a new room with a given name
  * @param { string } name 
  */
@@ -135,9 +101,35 @@ Room.updateRoomName = async function(id, name)
 	return res;
 }
 
-Room.addItem = function(){}
-Room.removeItem = function(){}
+/**
+ * Add an item (generated by the Item model schema) to a given room
+ * @param { string } id 
+ * @param { JSON } item 
+ */
+Room.addItem = async function(id, item)
+{
+	const itemId = uuidv4();
+
+	let room = await rethinkdb.db("dd_data").table("rooms").get(id).run(database.connection);
+
+	let items = room.items || [];
+	item.id = itemId;
+	items.push(item);
+
+	await Room.updateProperty(id, { "items": items });
+}
+
+/**
+ * Clear the items of a given room
+ * @param { string } id 
+ */
+Room.removeItem = async function(id)
+{
+	await Room.updateProperty(id, { "items": [] });
+}
+
 Room.clearItems = function(){}
+
 Room.editItem = function(){}
 
 module.exports = Room;
