+++ server/routes/room.js
@@ -1,4 +1,5 @@
 const { Router } = require('express');
+let Item = require('../models/item.model');
 let Room = require("../models/room.model");
 
 const router = Router();
